Add support for passing file argument
Doing this required introducing proper CLI argument parsing. This was so
namespaces could be distinguished from the files. I also introduced
non-positional parsing for the nrepl, mostly for conformance.

You can now provide either a file or stdin to the binary, and it will be
used by kibit when linting.

This also introduces support for Neomake, as it requires you to lint the
file in-place, instead of being passed the current contents of the
buffer via stdin.